+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sanakenno/shared
 
+## 1.7.0
+
 ## 1.6.0
 
 ### Minor Changes